requirements.txt
PyYAML==5.4.1
semantic-version==2.10.0